Introduced InterfaceObjects
Models now are implemented in Objects, naming convention "<ModelName>Object".
Most Services now return these ModelObjects instead of just Objects. This is achieved through the TransformObservable decorator that casts the incoming objects into the ModelObjects.
This was done so that incoming data from the API can come with an implemented interface of functions that shouldn't be defined by the API (e.g. to return the route of an article of a given dataset would have on the frontend)