Fixed routerLinks in RecentlyUpdated Article and Search Page
The hyperlinks given there previously there had a fundamental issue: They were trying to get the router by asking the BACKEND.
That was unacceptable. As such a system was implemented that transformed incoming data into objects that could determine their own routerUrl with getAbsoluteRouterUrl().
The function was added through the Transform decorator in every single service that grabbed specific kinds of articles.
The Article Object then got a function that could, based on the article_type, grab those objects, create one of them using its own data and use that objects getAbsoluteRouterUrl()
This way, the frontend gets to dictate where its articles route to.
Not the backend.