Fixed item-character-article cancel routing
The cancel button when creating an item for a character with the item-character-create route previously routed to "..".
That is an invalid path that has no route associated with it for item-character-create.
Thus all routes were associated with name and a module implemented with functions based on that. This module had a function that could determine whether the name of a route matches a given name-string.