Added wrapper to edit selected session entries
With this session-update wrapper users can now edit session objects, without requiring the admin view or anything.
Does not yet allow to set the date field, but that will be a feature in the future.

This feature is implemented as a formly wrapper. It applies a button after the select field and opens a modal with the form to submit an AJAX request to the backend server to update the entry.