Added way to get route based on route names
This is to avoid hard-coding URLs in the future for it is annoying. Thus all routes now are enforced to have a name via typing. This name is used by the written routeFilter module to grab the correct Route object and fill it with parameters as necessary before returning a string with a correctly mashed together path.