Implemented RouteGuarding
JWTTokens now come with an array that specifies the permissions of this particular user.
These permissions are used to determine whether a user should have access to a given route or not.
For this purpose all routes now define in their data attribute a "requiredPermissions" attribute that defines which permissions are required to activate a given route.
All possible permissions are specified in app.constants.ts