Changed some Token Service functions to static
To be able to use some of the functionality of TokenService without injecting it, I opted to make some of the functions static and giving them a private counterpart that calls them to avoid repetition while also not breaking the internal API. This breaks any usage of the API outside of the service. As JWT-Interceptor.ts was the only other user of these functions, that was fixed in there.