Added error forwarding and warnings
So far when an error happened, the page did  nothing. This commit implements behaviour for when an error occurs. Service calls that were required to load a page can now be routed to the error page appropriate for the error. Service calls that are minor inconveniences now throw a warning.