Added FieldValidation against Special Characters
Any field that is used to generate a url is no longer allowed to have special characters, such as "?" as they cause issues in hyperlinking to them (e.g. "Maribel ?" created non-functional URLs).
This commit creates a formly validator and validation-error-message and applies them to the various form-fields (mostly names of characters, creatures, organizations etc.) that are URL-critical.