Removed the ability to create encounters for a given character
With the coming feature of diaryentries being made up of encounters, every encounters needs to be associated with a diarynetry.
That means the information where an encounter in a diaryentry should be placed is vital, but is confusing to communicate in the context of a character that you create an encounter for.
Thus the feature was bombed.