Created ways to handle encounters for a diaryentry
Diaryentries from now on are lists of encounters. These are the components to handle this, first a component to handle lists of encounters, and a second one to handle individual encounters.

This came with some adjustments to the diaryentry retrieval service as they now come with loads of encounters and encounter connections, which also needed to be adjusted in their models.