Changed JWT interceptor to ignore requests to specific URLs
Some API Endpoint URLs just do not require authentication (e.g. when asking for a password reset). In these cases, the Interceptor shall not act and just hand the request onwards. This change adds a list that just contains the endings of all URLs that shall be ignored for authentication