Changed TransformObservable and TransformArrayObservable
These decorators previously could only be used if a parameter was directly provided to them, which was a class. In order to be able to have generic services to inherit from, a generic parameter needs to be provided here. However, the specific class is not known to the generic service, so it must provide a variable, aka the generic parameter earlier mentioned. Decorators however can't access in their declaration the object context. However, inside of the decorator you have the "target" parameter, which is the object within which the decorator is being called. You can access that objects properties from there. If you thus put the class in a property of the object within which the decorator is called, you can then give the decorator just a string with the property name and call target['propertyName'] to access the property and thus get the class.