Changed transformation decorators to allow inheritance
Previously the decorators required the class to be passed directly as an argument into the decorator.
When using inheritance, this is not possible, as the decorators will run before the parent-service knows what that class will be. What you can do is give them the variable name on the service that WILL contain the class.
In the anonymous function you put into descriptor.value you can then access "this", which in that context is the generic service which, when this is being called, will have the property whose string you now have defined with the class you need. You can thus go this[modelClass] to access that property and the class you need.