Background image is now handled via the background image component
The idea is that the component fetches the URLs to all campaign images at the start upon first load. Then, when you call a specific page, that page writes the currently visited campaign into the global url params service. Background image has a subscription to that, so when an update occurs there, it'll swap images with a fade in fade out animation.