Refined when to block updating the currently selected campaign
The currently selected campaign now does not get updated when that would not be a change of value (which is as before) OR when the campaign set has not yet been loaded.

This is because "updatecurrentlyselectedcampaign" may be called when the service is still being initialized for the first time, during which you have to wait for the first call to go out.