Reworked the way permissions work and adjusted invoking  parent classes accordingly
From now on, permissions are checked against a user's role. On initialization, the campaign name in the URL is checked against the role for this campaign in the user's token. If the role is of appropriate grade to pass on a create/view/update/delete permission, this is set accordingly. No role = no access, unless you're admin/superuser. In such cases, you receive any permission necessary