FIX border entries not getting pushed to the bottom as they should
margin-top auto rule on the footer block should've pushed it to the bottom, but it didn't in mobile view.
That was because in mobile view the swiped-right animation set display to block, not flex.
That somehow lead to margin-top auto not pushing things to the bottom.
Setting it to flex fixed that.

However it enabled the individual list items to get squeezed together on screens
with insufficient vertical length.
For those, flex-shink:0 blocked that shrinkage.

Lastly a border was established to distinguish the bottom entries more
clearly from general navigation entries.